requirements.txt
Django>=2.2,<3.0
selenium>=3.141.0
requests>=2.25.0
python-Levenshtein>=0.12.0
python-docx>=0.8.11
Pillow>=8.0.0 